requirements.txt
amqp==5.1.1
arabic-reshaper==2.1.3
argcomplete==3.1.2
asgiref==3.4.1
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
boto3==1.28.70
botocore==1.31.70
Brotli==1.1.0
certifi==2021.10.8
cffi==1.16.0
cfn-flip==1.3.0
chardet==5.2.0
charset-normalizer==2.0.10
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
cryptography==41.0.5
cssselect2==0.7.0
defusedxml==0.7.1
distro==1.8.0
Django==4.0.1
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-sslserver==0.22
django-storages==1.14.2
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
durationpy==0.5
et-xmlfile==1.1.0
eventlet==0.36.1
fillpdf==0.7.1
fonttools==4.43.1
future==0.18.2
greenlet==3.0.3
gunicorn==23.0.0
hjson==3.1.0
html5lib==1.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
jmespath==1.0.1
JPype1==1.4.1
kappa==0.6.0
kombu==5.2.4
MarkupSafe==2.1.3
numpy==1.26.4
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.2
pathspec==0.11.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
pdfrw2==0.5.0
Pillow==10.1.0
placebo==0.9.0
prompt-toolkit==3.0.38
psutil==6.0.0
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.8.0
PyJWT==2.8.0
PyMuPDF==1.24.9
PyPDF2==1.26.0
pypdfium2==4.22.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2021.3
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.1
redis==4.5.1
reportlab==3.6.13
requests==2.27.1
s3transfer==0.7.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tabula-py==2.9.0
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tqdm==4.66.1
troposphere==4.5.0
typing_extensions==4.9.0
tzdata>=2024.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
Wand==0.6.11
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.1
xhtml2pdf==0.2.5
zopfli==0.2.3
python-decouple
